refactor(api): extract zip package handling to dedicated type
