+++ api/pkg/import/handler.go
@@ -1,14 +1,11 @@
 package _import
 
 import (
-	"archive/zip"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"net/http"
 	"os"
-	"path"
-	"strings"
 
 	"github.com/go-openapi/runtime/middleware"
 
@@ -65,27 +62,35 @@ func (ih *ImportHandler) HandleImport(
 		return import_operations.NewImportNotFound().WithPayload(&error)
 	}
 
-	file, err := ioutil.TempFile(ih.tempStorageDir, "importConfig*"+defaultImportArchiveExtension)
-	if err != nil {
-		return import_operations.NewImportBadRequest()
-	}
-
-	defer func() {
-		errDefer := file.Close()
-		if errDefer != nil {
-			logger.Warnf("Error closing temporary import archive %s: %v", file.Name(), errDefer)
+	tempFileName, err := func() (string, error) {
+		file, err := ioutil.TempFile(ih.tempStorageDir, "importConfig*"+defaultImportArchiveExtension)
+		if err != nil {
+			return "", err
 		}
 
-		errDefer = os.Remove(file.Name())
-		if errDefer != nil {
-			logger.Errorf("Error deleting temporary import archive %s: %v", file.Name(), errDefer)
-		}
+		defer func() {
+			errDefer := file.Close()
+			if errDefer != nil {
+				logger.Warnf("Error closing temporary import archive %s: %v", file.Name(), errDefer)
+			}
+		}()
+
+		_, err = io.Copy(file, params.ConfigPackage)
+		return file.Name(), err
 	}()
 
-	tempFileSize, err := io.Copy(file, params.ConfigPackage)
+	if tempFileName != "" {
+		defer func() {
+			errDefer := os.Remove(tempFileName)
+			if errDefer != nil {
+				logger.Errorf("Error deleting temporary import archive %s: %v", tempFileName, errDefer)
+			}
+		}()
+	}
+
 	if err != nil {
 		logger.Errorf("Error saving import archive: %v", err)
-		message := fmt.Sprintf("Error reading import archive: %s", err)
+		message := fmt.Sprintf("Error saving import archive: %s", err)
 		error := models.Error{
 			Code:    http.StatusBadRequest,
 			Message: &message,
@@ -93,9 +98,10 @@ func (ih *ImportHandler) HandleImport(
 		return import_operations.NewImportBadRequest().WithPayload(&error)
 	}
 
-	zipReader, err := zip.NewReader(file, tempFileSize)
+	m, err := NewManifest(tempFileName)
+
 	if err != nil {
-		logger.Errorf("Error opening import archive %s: %v", file.Name(), err)
+		logger.Errorf("Error opening import archive: %v", err)
 		message := fmt.Sprintf("Error opening import archive: %s", err)
 		error := models.Error{
 			Code:    http.StatusUnsupportedMediaType,
@@ -104,87 +110,35 @@ func (ih *ImportHandler) HandleImport(
 		return import_operations.NewImportUnsupportedMediaType().WithPayload(&error)
 	}
 
-	extractionDir := strings.TrimSuffix(file.Name(), defaultImportArchiveExtension)
-	err = os.Mkdir(extractionDir, os.ModeDir|os.ModePerm)
-	if err != nil {
-		logger.Errorf("Error creating folder %s for zip extraction: %v", extractionDir, err)
-		message := fmt.Sprintf("Error reading import archive: %s", err)
-		error := models.Error{
-			Code:    http.StatusBadRequest,
-			Message: &message,
-		}
-		return import_operations.NewImportBadRequest().WithPayload(&error)
-	}
+	// if err != nil {
+	// 	logger.Errorf("Error creating folder %s for zip extraction: %v", extractionDir, err)
+	// 	message := fmt.Sprintf("Error reading import archive: %s", err)
+	// 	error := models.Error{
+	// 		Code:    http.StatusBadRequest,
+	// 		Message: &message,
+	// 	}
+	// 	return import_operations.NewImportBadRequest().WithPayload(&error)
+	// }
 
 	defer func() {
-		errCleanup := os.RemoveAll(extractionDir)
-		if errCleanup != nil {
-			logger.Warnf("Error cleaning up extraction directory %s: %v", extractionDir, errCleanup)
+		manifestCloseErr := m.Close()
+		if manifestCloseErr != nil {
+			logger.Warnf("Error closing manifest %+v: %s", m, manifestCloseErr)
 		}
 	}()
 
-	err = extractZipFile(zipReader, extractionDir)
-
-	if err != nil {
-		logger.Errorf("Error extracting zip %s: %v", file.Name(), err)
-		message := fmt.Sprintf("Error extracting archive: %s", err)
-		error := models.Error{
-			Code:    http.StatusBadRequest,
-			Message: &message,
-		}
-		// TODO check if we want to return something != 400
-		return import_operations.NewImportBadRequest().WithPayload(&error)
-	}
+	// if err != nil {
+	// 	logger.Errorf("Error extracting zip %s: %v", , err)
+	// 	message := fmt.Sprintf("Error extracting archive: %s", err)
+	// 	error := models.Error{
+	// 		Code:    http.StatusBadRequest,
+	// 		Message: &message,
+	// 	}
+	// 	// TODO check if we want to return something != 400
+	// 	return import_operations.NewImportBadRequest().WithPayload(&error)
+	// }
 
 	// TODO validate and start mainfest import
 
 	return import_operations.NewImportOK()
-}
-
-func extractZipFile(reader *zip.Reader, outputDir string) error {
-
-	// TODO add constraint on maximum extracted archive size
-
-	for _, zippedFile := range reader.File {
-		logger.Debugf("Extracting file %+v to %s...", reader, outputDir)
-
-		if zippedFile.FileInfo().IsDir() {
-			fullOuputDirectoryName := path.Join(outputDir, zippedFile.Name)
-			err := os.Mkdir(fullOuputDirectoryName, 0755)
-			if err != nil {
-				return fmt.Errorf("error creating directory %s: %w", fullOuputDirectoryName, err)
-			}
-		} else {
-			err := func() error {
-				src, err := zippedFile.Open()
-				if err != nil {
-					return fmt.Errorf("error reading %s from archive: %w", zippedFile.Name, err)
-				}
-				defer src.Close()
-
-				dst, err := os.Create(path.Join(outputDir, zippedFile.Name))
-				if err != nil {
-					return fmt.Errorf(
-						"error creating file %s in output directory %s: %w", zippedFile.Name, outputDir, err,
-					)
-				}
-				defer dst.Close()
-
-				// TODO add safety checks on the actual written file size
-				_, err = io.Copy(dst, src)
-
-				if err != nil {
-					return fmt.Errorf("error extracting %s from archive into %s: %w", zippedFile.Name, dst.Name(), err)
-				}
-
-				return nil
-			}()
-
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
 }
